Refactor architecture to have Settings which is the data structure and SettingsHandler which makes HTTP calls
